masivelly copy webui related (but not only) code from master

+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/window/model/DocumentCollection.java
@@ -30,13 +30,14 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import de.metas.copy_with_details.CopyRecordRequest;
 import de.metas.copy_with_details.CopyRecordService;
-import de.metas.document.references.zoom_into.RecordWindowFinder;
+import de.metas.copy_with_details.CopyRecordRequest;
+import de.metas.copy_with_details.CopyRecordService;
 import de.metas.i18n.AdMessageKey;
+import de.metas.i18n.BooleanWithReason;
 import de.metas.letters.model.MADBoilerPlate;
 import de.metas.letters.model.MADBoilerPlate.BoilerPlateContext;
 import de.metas.letters.model.MADBoilerPlate.SourceDocument;
 import de.metas.logging.LogManager;
-import de.metas.ui.web.exceptions.EntityNotFoundException;
 import de.metas.ui.web.session.UserSession;
 import de.metas.ui.web.window.WindowConstants;
 import de.metas.ui.web.window.controller.DocumentPermissionsHelper;
@@ -56,12 +57,10 @@ import de.metas.ui.web.window.exceptions.InvalidDocumentPathException;
 import de.metas.ui.web.window.invalidation.DocumentToInvalidate;
 import de.metas.ui.web.window.invalidation.IncludedDocumentToInvalidate;
 import de.metas.ui.web.window.model.Document.CopyMode;
-import de.metas.ui.web.window.model.lookup.DocumentZoomIntoInfo;
 import de.metas.util.Check;
 import de.metas.util.Services;
 import lombok.AllArgsConstructor;
 import lombok.NonNull;
-import org.adempiere.ad.element.api.AdWindowId;
 import org.adempiere.ad.expression.api.IExpressionEvaluator.OnVariableNotFound;
 import org.adempiere.ad.expression.api.ILogicExpression;
 import org.adempiere.ad.expression.api.LogicExpressionResult;
@@ -98,6 +97,7 @@ public class DocumentCollection
 
 	private static final Logger logger = LogManager.getLogger(DocumentCollection.class);
 	public static final AdMessageKey MSG_CLONING_NOT_ALLOWED_FOR_CURRENT_WINDOW = AdMessageKey.of("de.metas.ui.web.window.model.DocumentCollection.CloningNotAllowedForCurrentWindow");
+	public static final AdMessageKey MSG_CREATE_NOT_ALLOWED = AdMessageKey.of(("de.metas.ui.web.window.model.DocumentCollection.CreateNotAllowed"));
 
 	private final DocumentDescriptorFactory documentDescriptorFactory;
 	private final UserSession userSession;
@@ -137,7 +137,8 @@ public class DocumentCollection
 	/**
 	 * Delegates to the {@link DocumentDescriptorFactory#isWindowIdSupported(WindowId)} of this instance's {@code documentDescriptorFactory}.
 	 */
-	public boolean isWindowIdSupported(@Nullable final WindowId windowId)
+	public boolean isWindowIdSupported(
+			@Nullable final WindowId windowId)
 	{
 		return documentDescriptorFactory.isWindowIdSupported(windowId);
 	}
@@ -171,12 +172,15 @@ public class DocumentCollection
 		return windowIds != null && !windowIds.isEmpty() ? ImmutableSet.copyOf(windowIds) : ImmutableSet.of();
 	}
 
-	public Document getDocumentReadonly(@NonNull final DocumentPath documentPath)
+	public Document getDocumentReadonly(
+			@NonNull final DocumentPath documentPath)
 	{
 		return forDocumentReadonly(documentPath, Function.identity());
 	}
 
-	public <R> R forDocumentReadonly(@NonNull final DocumentPath documentPath, @NonNull final Function<Document, R> documentProcessor)
+	public <R> R forDocumentReadonly(
+			@NonNull final DocumentPath documentPath,
+			@NonNull final Function<Document, R> documentProcessor)
 	{
 		final DocumentPath rootDocumentPath = documentPath.getRootDocumentPath();
 
@@ -200,7 +204,8 @@ public class DocumentCollection
 		});
 	}
 
-	private Document getOrLoadDocument(@NonNull final DocumentKey documentKey)
+	private Document getOrLoadDocument(
+			@NonNull final DocumentKey documentKey)
 	{
 		try
 		{
@@ -219,7 +224,8 @@ public class DocumentCollection
 		}
 	}
 
-	public <R> R forRootDocumentReadonly(@NonNull final DocumentPath documentPath, final Function<Document, R> rootDocumentProcessor)
+	public <R> R forRootDocumentReadonly(
+			@NonNull final DocumentPath documentPath, final Function<Document, R> rootDocumentProcessor)
 	{
 		final DocumentKey rootDocumentKey = DocumentKey.ofRootDocumentPath(documentPath.getRootDocumentPath());
 
@@ -353,14 +359,15 @@ public class DocumentCollection
 		final LogicExpressionResult allow = allowExpr.evaluateToResult(userSession.toEvaluatee(), OnVariableNotFound.ReturnNoResult);
 		if (allow.isFalse())
 		{
-			throw new AdempiereException("Create not allowed");
+			throw new AdempiereException(MSG_CREATE_NOT_ALLOWED);
 		}
 	}
 
 	/**
 	 * Retrieves document from repository
 	 */
-	private Document retrieveRootDocumentFromRepository(final DocumentKey documentKey)
+	private Document retrieveRootDocumentFromRepository(
+			@NonNull final DocumentKey documentKey)
 	{
 		final DocumentEntityDescriptor entityDescriptor = getDocumentEntityDescriptor(documentKey.getWindowId());
 
@@ -421,7 +428,8 @@ public class DocumentCollection
 		return result;
 	}
 
-	private void commitRootDocument(@NonNull final Document rootDocument)
+	private void commitRootDocument(
+			@NonNull final Document rootDocument)
 	{
 		Preconditions.checkState(rootDocument.isRootDocument(), "{} is not a root document", rootDocument);
 
@@ -444,17 +452,12 @@ public class DocumentCollection
 
 		//
 		// Make sure all events were collected for the case when we just created the new document
-		// FIXME: this is a workaround and in case we find out all events were collected, we just need to remove this.
+		//
+		// IMPORTANT: This case happens when the document is not yet saved, but it's not the first time we are patching it.
+		// e.g. we have a document with multiple mandatory fields, user is filling them one by one, after each change a PATCH is sent
 		if (wasNew)
 		{
-			logger.debug("Checking if we collected all events for the new document");
-			final Set<String> collectedFieldNames = rootDocument.getChangesCollector().collectFrom(rootDocument, () -> "new document, initially missed");
-			if (!collectedFieldNames.isEmpty())
-			{
-				logger.warn("We would expect all events to be auto-magically collected but it seems that not all of them were collected!"
-						+ "\n Missed (but collected now) field names were: {}" //
-						+ "\n Document path: {}", collectedFieldNames, rootDocument.getDocumentPath());
-			}
+			rootDocument.getChangesCollector().collectFrom(rootDocument, () -> "new document, initially missed");
 		}
 
 	}
@@ -474,8 +477,16 @@ public class DocumentCollection
 		}
 
 		forRootDocumentWritable(rootDocumentPath, changesCollector, rootDocument -> {
+
 			if (documentPath.isRootDocument())
 			{
+				final BooleanWithReason isDeleteForbidden = rootDocument.isDeleteForbidden();
+				if (isDeleteForbidden.isTrue())
+				{
+					throw new AdempiereException(isDeleteForbidden.getReason())
+							.markAsUserValidationError();
+				}
+
 				if (!rootDocument.isNew())
 				{
 					rootDocument.deleteFromRepository();
@@ -496,7 +507,8 @@ public class DocumentCollection
 		});
 	}
 
-	private void assertDeleteDocumentAllowed(final DocumentEntityDescriptor entityDescriptor)
+	private void assertDeleteDocumentAllowed(
+			@NonNull final DocumentEntityDescriptor entityDescriptor)
 	{
 		final Evaluatee evalCtx = Evaluatees.mapBuilder()
 				.put(WindowConstants.FIELDNAME_Processed, false)
@@ -525,98 +537,11 @@ public class DocumentCollection
 		return documentDescriptorFactory.getTableRecordReference(documentPath);
 	}
 
-	public WindowId getWindowId(@NonNull final DocumentZoomIntoInfo zoomIntoInfo)
-	{
-		if (zoomIntoInfo.getWindowId() != null)
-		{
-			return zoomIntoInfo.getWindowId();
-		}
-
-		final AdWindowId zoomInto_adWindowId;
-		if (zoomIntoInfo.isRecordIdPresent())
-		{
-			zoomInto_adWindowId = RecordWindowFinder.newInstance(zoomIntoInfo.getTableName(), zoomIntoInfo.getRecordId())
-					.checkRecordPresentInWindow()
-					.checkParentRecord()
-					.findAdWindowId()
-					.orElse(null);
-		}
-		else
-		{
-			zoomInto_adWindowId = RecordWindowFinder.findAdWindowId(zoomIntoInfo.getTableName()).orElse(null);
-		}
-
-		if (zoomInto_adWindowId == null)
-		{
-			throw new EntityNotFoundException("No windowId found")
-					.setParameter("zoomIntoInfo", zoomIntoInfo);
-		}
-
-		return WindowId.of(zoomInto_adWindowId);
-	}
-
 	public boolean isValidDocumentPath(final DocumentPath documentPath)
 	{
 		return documentPath != null
 				&& documentPath.getWindowId().isInt()
 				&& documentPath.getDocumentId().isInt();
-	}
-
-	/**
-	 * Retrieves document path for given ZoomInto info.
-	 */
-	public DocumentPath getDocumentPath(@NonNull final DocumentZoomIntoInfo zoomIntoInfo)
-	{
-		if (!zoomIntoInfo.isRecordIdPresent())
-		{
-			throw new IllegalArgumentException("recordId must be set in " + zoomIntoInfo);
-		}
-
-		//
-		// Find the root window ID
-		final WindowId zoomIntoWindowIdEffective = getWindowId(zoomIntoInfo);
-
-		final DocumentEntityDescriptor rootEntityDescriptor = getDocumentEntityDescriptor(zoomIntoWindowIdEffective);
-		final String zoomIntoTableName = zoomIntoInfo.getTableName();
-
-		//
-		// We are dealing with a root document
-		// (i.e. root descriptor's table is matching record's table)
-		if (Objects.equals(rootEntityDescriptor.getTableName(), zoomIntoTableName))
-		{
-			final DocumentId rootDocumentId = DocumentId.of(zoomIntoInfo.getRecordId());
-			return DocumentPath.rootDocumentPath(zoomIntoWindowIdEffective, rootDocumentId);
-		}
-		//
-		// We are dealing with an included document
-		else
-		{
-			// Search the root descriptor for any child entity descriptor which would match record's TableName
-			final List<DocumentEntityDescriptor> childEntityDescriptors = rootEntityDescriptor.getIncludedEntities().stream()
-					.filter(includedEntityDescriptor -> Objects.equals(includedEntityDescriptor.getTableName(), zoomIntoTableName))
-					.collect(ImmutableList.toImmutableList());
-			if (childEntityDescriptors.isEmpty())
-			{
-				throw new EntityNotFoundException("Cannot find the detail tab to zoom into")
-						.setParameter("zoomIntoInfo", zoomIntoInfo)
-						.setParameter("zoomIntoWindowId", zoomIntoWindowIdEffective)
-						.setParameter("rootEntityDescriptor", rootEntityDescriptor);
-			}
-			else if (childEntityDescriptors.size() > 1)
-			{
-				logger.warn("More then one child descriptors matched our root descriptor. Picking the fist one. \nRoot descriptor: {} \nChild descriptors: {}", rootEntityDescriptor, childEntityDescriptors);
-			}
-			//
-			final DocumentEntityDescriptor childEntityDescriptor = childEntityDescriptors.get(0);
-
-			// Find the root DocumentId
-			final DocumentId rowId = DocumentId.of(zoomIntoInfo.getRecordId());
-			final DocumentId rootDocumentId = DocumentQuery.ofRecordId(childEntityDescriptor, rowId)
-					.retrieveParentDocumentId(rootEntityDescriptor);
-
-			//
-			return DocumentPath.includedDocumentPath(zoomIntoWindowIdEffective, rootDocumentId, Check.assumeNotNull(childEntityDescriptor.getDetailId(), "Expected childEntityDescriptor.getDetailId not null"), rowId);
-		}
 	}
 
 	public DocumentWebsocketPublisher getWebsocketPublisher()
@@ -653,7 +578,8 @@ public class DocumentCollection
 		documentKeys.forEach(documentKey -> websocketPublisher.staleRootDocument(documentKey.getWindowId(), documentKey.getDocumentId(), true));
 	}
 
-	public void invalidateDocumentsByWindowId(@NonNull final WindowId windowId)
+	public void invalidateDocumentsByWindowId(
+			@NonNull final WindowId windowId)
 
 	{
 		final ImmutableList<DocumentKey> documentKeys = rootDocuments.asMap().keySet()
@@ -669,10 +595,13 @@ public class DocumentCollection
 
 	public void invalidateAll(final Collection<DocumentToInvalidate> documentToInvalidateList)
 	{
-		documentToInvalidateList.forEach(this::invalidate);
+		for (final DocumentToInvalidate documentToInvalidate : documentToInvalidateList)
+		{
+			invalidate(documentToInvalidate);
+		}
 	}
 
-	private void invalidate(final DocumentToInvalidate documentToInvalidate)
+	private void invalidate(@NonNull final DocumentToInvalidate documentToInvalidate)
 	{
 		final ImmutableList<DocumentEntityDescriptor> entityDescriptors = getCachedWindowIdsForTableName(documentToInvalidate.getTableName())
 				.stream()
@@ -692,31 +621,40 @@ public class DocumentCollection
 			final Document rootDocument = rootDocuments.getIfPresent(rootDocumentKey);
 			if (rootDocument != null)
 			{
-				try (final IAutoCloseable ignored = rootDocument.lockForWriting())
+				//
+				// Invalidate included documents
+				// NOTE: we do this even if we will have to invalidate the whole document because we want to collect the events for frontend.
+				// Ideally would be to just invalidate the root document if that was required and frontend had to deal with it.
+				final Collection<IncludedDocumentToInvalidate> includedDocumentsToInvalidate = documentToInvalidate.getIncludedDocuments();
+				if (!includedDocumentsToInvalidate.isEmpty())
 				{
-					for (final IncludedDocumentToInvalidate includedDocumentToInvalidate : documentToInvalidate.getIncludedDocuments())
+					try (final IAutoCloseable ignored = rootDocument.lockForWriting())
 					{
-						final DocumentIdsSelection includedRowIds = includedDocumentToInvalidate.toDocumentIdsSelection();
-						if (includedRowIds.isEmpty())
+						for (final IncludedDocumentToInvalidate includedDocumentToInvalidate : includedDocumentsToInvalidate)
 						{
-							continue;
-						}
+							final DocumentIdsSelection includedRowIds = includedDocumentToInvalidate.toDocumentIdsSelection();
+							if (includedRowIds.isEmpty())
+							{
+								continue;
+							}
 
-						for (final DocumentEntityDescriptor includedEntityDescriptor : entityDescriptor.getIncludedEntitiesByTableName(includedDocumentToInvalidate.getTableName()))
-						{
-							final DetailId detailId = includedEntityDescriptor.getDetailId();
-
-							rootDocument.getIncludedDocumentsCollection(Check.assumeNotNull(detailId, "Expected detailId not null")).markStale(includedRowIds);
+							for (final DocumentEntityDescriptor includedEntityDescriptor : entityDescriptor.getIncludedEntitiesByTableName(includedDocumentToInvalidate.getTableName()))
+							{
+								final DetailId detailId = Check.assumeNotNull(includedEntityDescriptor.getDetailId(), "Expected detailId not null");
+								rootDocument.getIncludedDocumentsCollection(detailId).markStale(includedRowIds);
+							}
 						}
 					}
 				}
-			}
 
-			//
-			// Invalidate the root document
-			if (documentToInvalidate.isInvalidateDocument())
-			{
-				rootDocuments.invalidate(rootDocumentKey);
+				//
+				// Invalidate the root document
+				// NOTE: avoid invalidating if the document is new (and not saved) because in that case we will lose the document and we will never be able to recover.
+				// As a symptom the user will get 404 or similar in his browser and the document will vanish completely.
+				if (documentToInvalidate.isInvalidateDocument() && !rootDocument.isNew())
+				{
+					rootDocuments.invalidate(rootDocumentKey);
+				}
 			}
 
 			//
@@ -753,7 +691,8 @@ public class DocumentCollection
 	/**
 	 * Invalidates all root documents identified by tableName/recordId and notifies frontend (via websocket).
 	 */
-	public void invalidateRootDocument(@NonNull final DocumentPath documentPath)
+	public void invalidateRootDocument(
+			@NonNull final DocumentPath documentPath)
 	{
 		final DocumentKey documentKey = DocumentKey.ofRootDocumentPath(documentPath);
 
@@ -845,13 +784,13 @@ public class DocumentCollection
 	@Immutable
 	private static final class DocumentKey
 	{
-		public static DocumentKey of(final Document document)
+		public static DocumentKey of(@NonNull final Document document)
 		{
 			final DocumentPath documentPath = document.getDocumentPath();
 			return ofRootDocumentPath(documentPath);
 		}
 
-		public static DocumentKey ofRootDocumentPath(final DocumentPath documentPath)
+		public static DocumentKey ofRootDocumentPath(@NonNull final DocumentPath documentPath)
 		{
 			if (!documentPath.isRootDocument())
 			{
@@ -861,10 +800,16 @@ public class DocumentCollection
 			{
 				throw new InvalidDocumentPathException(documentPath, "document path for creating new documents is not allowed");
 			}
-			return new DocumentKey(documentPath.getDocumentType(), documentPath.getDocumentTypeId(), documentPath.getDocumentId());
+
+			return new DocumentKey(
+					documentPath.getDocumentType(),
+					documentPath.getDocumentTypeId(),
+					documentPath.getDocumentId());
 		}
 
-		public static DocumentKey of(@NonNull final WindowId windowId, @NonNull final DocumentId documentId)
+		public static DocumentKey of(
+				@NonNull final WindowId windowId,
+				@NonNull final DocumentId documentId)
 		{
 			return new DocumentKey(DocumentType.Window, windowId.toDocumentId(), documentId);
 		}
@@ -875,11 +820,14 @@ public class DocumentCollection
 
 		private Integer _hashcode = null;
 
-		private DocumentKey(final DocumentType documentType, final DocumentId documentTypeId, final DocumentId documentId)
+		private DocumentKey(
+				@NonNull final DocumentType documentType,
+				@NonNull final DocumentId documentTypeId,
+				@NonNull final DocumentId documentId)
 		{
-			this.documentType = Preconditions.checkNotNull(documentType, "documentType");
-			this.documentTypeId = Preconditions.checkNotNull(documentTypeId, "documentTypeId");
-			this.documentId = Preconditions.checkNotNull(documentId, "documentId");
+			this.documentType = documentType;
+			this.documentTypeId = documentTypeId;
+			this.documentId = documentId;
 		}
 
 		@Override
